Refactor: moved delivery getters in DeliveryQuery
