+++ libs/material/src/lib/delivery/+state/delivery.service.ts
@@ -16,14 +16,12 @@ import { TemplateQuery } from '../../template/+state';
 })
 export class DeliveryService {
   constructor(
-    private firestore: AngularFirestore,
     private movieQuery: MovieQuery,
     private organizationQuery: OrganizationQuery,
     private materialQuery: MaterialQuery,
     private query: DeliveryQuery,
     private store: DeliveryStore,
     private templateQuery: TemplateQuery,
-    private materialStore: MaterialStore,
     private db: AngularFirestore
   ) {}
 
@@ -44,117 +42,19 @@ export class DeliveryService {
 
   public deliveredToggle(material: Material, movieId: string) {
     // Change material 'delivered' property value to true or false when triggered
-    return this.firestore
+    return this.db
       .doc<Material>(`movies/${movieId}/materials/${material.id}`)
       .update({ delivered: !material.delivered });
   }
 
-  public get deliveryMaterialsByActiveMovie$() {
-    // Get and sort the active movie's delivery's materials by category
-    return this.movieQuery.selectActive().pipe(
-      filter(movie => !!movie),
-      switchMap(movie =>
-        this.firestore.collection<Material>(`movies/${movie.id}/materials`).valueChanges()
-      ),
-      map(materials => materialsByCategory(materials))
-    );
-  }
-
-  public get deliveriesByActiveMovie$() {
-    // Get a list of deliveries for the active movie
-    return this.movieQuery.selectActiveId().pipe(
-      filter(id => !!id),
-      switchMap(id =>
-        this.firestore
-          .collection<Delivery>('deliveries', ref => ref.where('movieId', '==', id))
-          .valueChanges()
-      ),
-      tap(deliveries => this.store.set(deliveries))
-    );
-  }
-
-  public get materialsByActiveDelivery$() {
-    // Get the movie materials for the active delivery (with 'delivered' property)
-    return this.movieQuery.selectActive().pipe(
-      filter(movie => !!movie),
-      switchMap(movie =>
-        this.firestore.collection<Material>(`movies/${movie.id}/materials`).valueChanges()
-      ),
-      tap(materials => this.materialStore.set(materials)),
-      map(materials => {
-        const id = this.query.getActiveId();
-        const deliveryMaterials = [];
-        materials.forEach(material => {
-          if (!!material && material.deliveriesIds.includes(id)) {
-            deliveryMaterials.push(material);
-          }
-        });
-        return deliveryMaterials;
-      })
-    );
-  }
-
-  public get sortedDeliveryMaterials$() {
-    // Sort the active delivery's materials by category
-    return this.materialsByActiveDelivery$.pipe(map(materials => materialsByCategory(materials)));
-  }
-
-  public get deliveryProgression$() {
-    // Get the progression % of the delivery
-    return this.movieQuery.selectActive().pipe(
-      filter(movie => !!movie),
-      switchMap(movie =>
-        this.firestore.collection<Material>(`movies/${movie.id}/materials`).valueChanges()
-      ),
-      tap(materials => this.materialStore.set(materials)),
-      map(materials => {
-        const id = this.query.getActiveId();
-        const deliveredMaterials = [];
-        const totalMaterials = [];
-        materials.forEach(material => {
-          if (!!material && material.deliveriesIds.includes(id)) {
-            totalMaterials.push(material);
-            if (material.delivered === true) {
-              deliveredMaterials.push(material);
-            }
-          }
-        });
-        return Math.round((deliveredMaterials.length / (totalMaterials.length / 100)) * 10) / 10;
-      })
-    );
-  }
-
-  public get movieProgression$() {
-    // Get the progression % of the movie's deliveries
-    return this.movieQuery.selectActive().pipe(
-      filter(movie => !!movie),
-      switchMap(movie =>
-        this.firestore.collection<Material>(`movies/${movie.id}/materials`).valueChanges()
-      ),
-      map(materials => {
-        const deliveredMaterials = [];
-        const totalMaterials = [];
-        materials.forEach(material => {
-          if (!!material) {
-            totalMaterials.push(material);
-            if (material.delivered === true) {
-              deliveredMaterials.push(material);
-            }
-          }
-        });
-        return Math.round((deliveredMaterials.length / (totalMaterials.length / 100)) * 10) / 10;
-      })
-    );
-  }
-
   public addDelivery(id: string, templateId?: string) {
     const orgId = this.organizationQuery.getActiveId();
-    const stakeholderId = this.firestore.createId();
+    const stakeholderId = this.db.createId();
     const delivery = createDelivery({ id, movieId: this.movieQuery.getActiveId() });
     const stakeholder = createStakeholder({ id: stakeholderId, orgId });
 
-    this.firestore.doc<Delivery>(`deliveries/${id}`).set(delivery);
-    this.firestore
+    this.db.doc<Delivery>(`deliveries/${id}`).set(delivery);
+    this.db
       .doc<Stakeholder>(`deliveries/${id}/stakeholders/${stakeholderId}`)
       .set(stakeholder);
     this.store.setActive(id);
